requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
